+++ app/controllers/general.php
@@ -274,11 +274,18 @@ App::shutdown(function ($utopia, $request, $response, $project, $webhooks, $audi
     /** @var bool $mode */
 
     if (!empty($functions->getParam('event'))) {
-        $functions->setParam('payload', $webhooks->getParam('payload'));
+        if(empty($functions->getParam('payload'))) {
+            $functions->setParam('payload', $response->getPayload());
+        }
+
         $functions->trigger();
     }
 
     if (!empty($webhooks->getParam('event'))) {
+        if(empty($webhooks->getParam('payload'))) {
+            $webhooks->setParam('payload', $response->getPayload());
+        }
+
         $webhooks->trigger();
     }